chore(docs): remove media-insert shortcut
