+++ shortcuts/doc/doc_errors_test.go
@@ -86,84 +86,6 @@ func assertValidationContract(t *testing.T, err error, wantSubtype errs.Subtype,
 	}
 }
 
-func TestDocMediaInsertValidateContract(t *testing.T) {
-	cases := []struct {
-		name       string
-		str        map[string]string
-		bools      map[string]bool
-		ints       map[string]int
-		wantParam  string
-		wantParams []string
-	}{
-		{
-			name:       "neither file nor clipboard",
-			str:        map[string]string{"doc": testDocxToken},
-			wantParam:  "", // one-of-two flags: enumerated in Params
-			wantParams: []string{"--file", "--from-clipboard"},
-		},
-		{
-			name:       "file and clipboard together",
-			str:        map[string]string{"doc": testDocxToken, "file": "dummy.png"},
-			bools:      map[string]bool{"from-clipboard": true},
-			wantParam:  "", // mutual exclusion: enumerated in Params
-			wantParams: []string{"--file", "--from-clipboard"},
-		},
-		{
-			name:      "non-docx document",
-			str:       map[string]string{"doc": "https://example.larksuite.com/doc/xxxxxx", "file": "dummy.png"},
-			wantParam: "--doc",
-		},
-		{
-			name:      "invalid file-view",
-			str:       map[string]string{"doc": testDocxToken, "file": "dummy.png", "file-view": "bogus"},
-			wantParam: "--file-view",
-		},
-		{
-			name:      "file-view without type file",
-			str:       map[string]string{"doc": testDocxToken, "file": "dummy.png", "file-view": "card", "type": "image"},
-			wantParam: "--file-view",
-		},
-		{
-			name:       "dimensions with non-image type",
-			str:        map[string]string{"doc": testDocxToken, "file": "dummy.png", "type": "file"},
-			ints:       map[string]int{"width": 100},
-			wantParam:  "", // only --width was set here, so only it is enumerated
-			wantParams: []string{"--width"},
-		},
-		{
-			name:      "non-positive width",
-			str:       map[string]string{"doc": testDocxToken, "file": "dummy.png", "type": "image"},
-			ints:      map[string]int{"width": 0},
-			wantParam: "--width",
-		},
-		{
-			name:      "non-positive height",
-			str:       map[string]string{"doc": testDocxToken, "file": "dummy.png", "type": "image"},
-			ints:      map[string]int{"height": 0},
-			wantParam: "--height",
-		},
-		{
-			name:      "width over maximum",
-			str:       map[string]string{"doc": testDocxToken, "file": "dummy.png", "type": "image"},
-			ints:      map[string]int{"width": 10001},
-			wantParam: "--width",
-		},
-		{
-			name:      "height over maximum",
-			str:       map[string]string{"doc": testDocxToken, "file": "dummy.png", "type": "image"},
-			ints:      map[string]int{"height": 10001},
-			wantParam: "--height",
-		},
-	}
-	for _, tc := range cases {
-		t.Run(tc.name, func(t *testing.T) {
-			rt := docValidateRuntime(t, tc.str, tc.bools, tc.ints)
-			err := DocMediaInsert.Validate(context.Background(), rt)
-			assertValidationContract(t, err, errs.SubtypeInvalidArgument, tc.wantParam, tc.wantParams...)
-		})
-	}
-}
-
 func TestValidateCreateV2Contract(t *testing.T) {
 	cases := []struct {
 		name       string
