+++ shortcuts/doc/doc_media_test.go
@@ -7,7 +7,6 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
-	"fmt"
 	"net/http"
 	"os"
 	"path/filepath"
@@ -79,103 +78,6 @@ func withDocsWorkingDir(t *testing.T, dir string) {
 			t.Fatalf("restore cwd error: %v", err)
 		}
 	})
-}
-
-func TestDocMediaInsertRejectsOldDocURL(t *testing.T) {
-	f, _, _, _ := cmdutil.TestFactory(t, docsTestConfigWithAppID("docs-test-app"))
-
-	err := mountAndRunDocs(t, DocMediaInsert, []string{
-		"+media-insert",
-		"--doc", "https://example.larksuite.com/doc/xxxxxx",
-		"--file", "dummy.png",
-		"--dry-run",
-		"--as", "bot",
-	}, f, nil)
-	if err == nil {
-		t.Fatal("expected validation error, got nil")
-	}
-	if !strings.Contains(err.Error(), "only supports docx documents") {
-		t.Fatalf("unexpected error: %v", err)
-	}
-}
-
-func TestDocMediaInsertValidateRequiresFileOrClipboard(t *testing.T) {
-	f, _, _, _ := cmdutil.TestFactory(t, docsTestConfigWithAppID("docs-test-app"))
-
-	err := mountAndRunDocs(t, DocMediaInsert, []string{
-		"+media-insert",
-		"--doc", "https://example.larksuite.com/docx/doxcnXXXXXXXXXXXXXXXXXX",
-		"--dry-run",
-		"--as", "bot",
-	}, f, nil)
-	if err == nil {
-		t.Fatal("expected validation error, got nil")
-	}
-	if !strings.Contains(err.Error(), "one of --file or --from-clipboard is required") {
-		t.Fatalf("unexpected error: %v", err)
-	}
-}
-
-func TestDocMediaInsertValidateRejectsFileAndClipboardTogether(t *testing.T) {
-	f, _, _, _ := cmdutil.TestFactory(t, docsTestConfigWithAppID("docs-test-app"))
-
-	err := mountAndRunDocs(t, DocMediaInsert, []string{
-		"+media-insert",
-		"--doc", "https://example.larksuite.com/docx/doxcnXXXXXXXXXXXXXXXXXX",
-		"--file", "dummy.png",
-		"--from-clipboard",
-		"--dry-run",
-		"--as", "bot",
-	}, f, nil)
-	if err == nil {
-		t.Fatal("expected mutual-exclusion error, got nil")
-	}
-	if !strings.Contains(err.Error(), "mutually exclusive") {
-		t.Fatalf("unexpected error: %v", err)
-	}
-}
-
-func TestDocMediaInsertDryRunWithClipboardUsesPlaceholder(t *testing.T) {
-	f, stdout, _, _ := cmdutil.TestFactory(t, docsTestConfigWithAppID("docs-test-app"))
-
-	err := mountAndRunDocs(t, DocMediaInsert, []string{
-		"+media-insert",
-		"--doc", "https://example.larksuite.com/docx/doxcnXXXXXXXXXXXXXXXXXX",
-		"--from-clipboard",
-		"--dry-run",
-		"--as", "bot",
-	}, f, stdout)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	// JSON output escapes "<" and ">" as \u003c / \u003e by default.
-	out := stdout.String()
-	if !strings.Contains(out, `\u003cclipboard image\u003e`) && !strings.Contains(out, "<clipboard image>") {
-		t.Fatalf("dry-run output missing <clipboard image> placeholder: %s", out)
-	}
-}
-
-func TestDocMediaInsertDryRunWikiAddsResolveStep(t *testing.T) {
-	f, stdout, _, _ := cmdutil.TestFactory(t, docsTestConfigWithAppID("docs-test-app"))
-
-	err := mountAndRunDocs(t, DocMediaInsert, []string{
-		"+media-insert",
-		"--doc", "https://example.larksuite.com/wiki/xxxxxx",
-		"--file", "dummy.png",
-		"--dry-run",
-		"--as", "bot",
-	}, f, stdout)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	out := stdout.String()
-	if !strings.Contains(out, "Resolve wiki node to docx document") {
-		t.Fatalf("dry-run output missing wiki resolve step: %s", out)
-	}
-	if !strings.Contains(out, "resolved_docx_token") {
-		t.Fatalf("dry-run output missing resolved docx token placeholder: %s", out)
-	}
 }
 
 func TestDocMediaParentType(t *testing.T) {
@@ -359,57 +261,6 @@ func TestDocMediaUploadDryRunUsesMultipartForLargeFile(t *testing.T) {
 	}
 }
 
-func TestDocMediaInsertDryRunUsesMultipartForLargeFile(t *testing.T) {
-	tmpDir := t.TempDir()
-	withDocsWorkingDir(t, tmpDir)
-	writeSizedDocTestFile(t, "large.bin", common.MaxDriveMediaUploadSinglePartSize+1)
-
-	cmd := &cobra.Command{Use: "docs +media-insert"}
-	cmd.Flags().String("file", "", "")
-	cmd.Flags().String("doc", "", "")
-	cmd.Flags().String("type", "", "")
-	cmd.Flags().String("align", "", "")
-	cmd.Flags().String("caption", "", "")
-	if err := cmd.Flags().Set("doc", "doxcnDryRunLarge"); err != nil {
-		t.Fatalf("set --doc: %v", err)
-	}
-	if err := cmd.Flags().Set("file", "./large.bin"); err != nil {
-		t.Fatalf("set --file: %v", err)
-	}
-
-	dry := decodeDocDryRun(t, DocMediaInsert.DryRun(context.Background(), common.TestNewRuntimeContext(cmd, nil)))
-	if dry.Description != "4-step orchestration: query root → create block → upload file → bind to block (auto-rollback on failure)" {
-		t.Fatalf("dry-run description = %q", dry.Description)
-	}
-	if len(dry.API) != 6 {
-		t.Fatalf("expected 6 API calls, got %d", len(dry.API))
-	}
-	if dry.API[2].URL != "/open-apis/drive/v1/medias/upload_prepare" {
-		t.Fatalf("third URL = %q, want upload_prepare", dry.API[2].URL)
-	}
-	if dry.API[3].URL != "/open-apis/drive/v1/medias/upload_part" {
-		t.Fatalf("fourth URL = %q, want upload_part", dry.API[3].URL)
-	}
-	if dry.API[4].URL != "/open-apis/drive/v1/medias/upload_finish" {
-		t.Fatalf("fifth URL = %q, want upload_finish", dry.API[4].URL)
-	}
-	if dry.API[5].URL != "/open-apis/docx/v1/documents/doxcnDryRunLarge/blocks/batch_update" {
-		t.Fatalf("last URL = %q, want batch_update", dry.API[5].URL)
-	}
-	if !strings.Contains(dry.API[2].Desc, "[3a]") {
-		t.Fatalf("upload_prepare desc = %q, want [3a] step marker", dry.API[2].Desc)
-	}
-	if !strings.Contains(dry.API[3].Desc, "[3b]") {
-		t.Fatalf("upload_part desc = %q, want [3b] step marker", dry.API[3].Desc)
-	}
-	if !strings.Contains(dry.API[4].Desc, "[3c]") {
-		t.Fatalf("upload_finish desc = %q, want [3c] step marker", dry.API[4].Desc)
-	}
-	if !strings.Contains(dry.API[5].Desc, "[4]") {
-		t.Fatalf("batch_update desc = %q, want [4] step marker", dry.API[5].Desc)
-	}
-}
-
 func TestUploadDocMediaFileWithContentUsesSinglePartUpload(t *testing.T) {
 	// Clipboard path: in-memory bytes (no FilePath) route through
 	// UploadDriveMediaAllTyped when small enough. This also exercises the
@@ -509,148 +360,6 @@ func TestUploadDocMediaFileWithContentUsesMultipart(t *testing.T) {
 	}
 	if fileToken != "file_content_multi_done" {
 		t.Fatalf("fileToken = %q, want %q", fileToken, "file_content_multi_done")
-	}
-}
-
-func TestDocMediaInsertExecuteFromClipboard(t *testing.T) {
-	// Covers the Execute clipboard branch end-to-end: read synthetic bytes,
-	// resolve docx root, create block, upload in-memory content, bind to block.
-	prev := readClipboardImage
-	t.Cleanup(func() { readClipboardImage = prev })
-	payload := []byte{0x89, 0x50, 0x4e, 0x47, 0x0d, 0x0a, 0x1a, 0x0a, 0xAA, 0xBB}
-	readClipboardImage = func() ([]byte, error) { return payload, nil }
-
-	f, stdout, stderr, reg := cmdutil.TestFactory(t, docsTestConfigWithAppID("docs-clipboard-exec-app"))
-	documentID := "doxcnClipboardExec1"
-
-	// Step 1: GET root block
-	reg.Register(&httpmock.Stub{
-		Method: "GET",
-		URL:    "/open-apis/docx/v1/documents/" + documentID + "/blocks/" + documentID,
-		Body: map[string]interface{}{
-			"code": 0, "msg": "ok",
-			"data": map[string]interface{}{
-				"block": map[string]interface{}{
-					"block_id": documentID,
-					"children": []interface{}{"existing_block"},
-				},
-			},
-		},
-	})
-	// Step 2: POST create child block
-	reg.Register(&httpmock.Stub{
-		Method: "POST",
-		URL:    "/open-apis/docx/v1/documents/" + documentID + "/blocks/" + documentID + "/children",
-		Body: map[string]interface{}{
-			"code": 0, "msg": "ok",
-			"data": map[string]interface{}{
-				"children": []interface{}{
-					map[string]interface{}{"block_id": "new_image_block"},
-				},
-			},
-		},
-	})
-	// Step 3: POST upload_all for in-memory bytes
-	uploadStub := &httpmock.Stub{
-		Method: "POST",
-		URL:    "/open-apis/drive/v1/medias/upload_all",
-		Body: map[string]interface{}{
-			"code": 0,
-			"data": map[string]interface{}{"file_token": "file_clip_abc"},
-		},
-	}
-	reg.Register(uploadStub)
-	// Step 4: PATCH batch_update
-	reg.Register(&httpmock.Stub{
-		Method: "PATCH",
-		URL:    "/open-apis/docx/v1/documents/" + documentID + "/blocks/batch_update",
-		Body:   map[string]interface{}{"code": 0, "msg": "ok"},
-	})
-
-	err := mountAndRunDocs(t, DocMediaInsert, []string{
-		"+media-insert",
-		"--doc", documentID,
-		"--from-clipboard",
-		"--as", "bot",
-	}, f, stdout)
-	if err != nil {
-		t.Fatalf("unexpected error: %v — stderr: %s", err, stderr.String())
-	}
-
-	// stderr should show clipboard read + file name "clipboard.png"
-	if !strings.Contains(stderr.String(), "Reading image from clipboard") {
-		t.Errorf("stderr missing clipboard-read log: %s", stderr.String())
-	}
-	if !strings.Contains(stderr.String(), "clipboard.png") {
-		t.Errorf("stderr missing clipboard.png file name: %s", stderr.String())
-	}
-	// stdout should include the file_token
-	if !strings.Contains(stdout.String(), "file_clip_abc") {
-		t.Errorf("stdout missing file_token: %s", stdout.String())
-	}
-
-	// Upload multipart body should contain the synthetic payload bytes.
-	if !bytes.Contains(uploadStub.CapturedBody, payload) {
-		t.Errorf("upload body missing clipboard payload bytes")
-	}
-}
-
-func TestDocMediaInsertExecuteClipboardReadError(t *testing.T) {
-	// Covers the early-return when clipboard read fails (no osascript etc).
-	prev := readClipboardImage
-	t.Cleanup(func() { readClipboardImage = prev })
-	readClipboardImage = func() ([]byte, error) {
-		return nil, fmt.Errorf("clipboard image upload is not supported on test")
-	}
-
-	f, _, _, _ := cmdutil.TestFactory(t, docsTestConfigWithAppID("docs-clipboard-err-app"))
-	err := mountAndRunDocs(t, DocMediaInsert, []string{
-		"+media-insert",
-		"--doc", "doxcnXXXXXXXXXXXXXXXXXX",
-		"--from-clipboard",
-		"--as", "bot",
-	}, f, nil)
-	if err == nil {
-		t.Fatal("expected clipboard read error, got nil")
-	}
-	if !strings.Contains(err.Error(), "clipboard image upload is not supported") {
-		t.Fatalf("unexpected error: %v", err)
-	}
-}
-
-func TestDocMediaInsertExecuteResolvesWikiBeforeFileCheck(t *testing.T) {
-	f, _, stderr, reg := cmdutil.TestFactory(t, docsTestConfigWithAppID("docs-insert-exec-app"))
-	reg.Register(&httpmock.Stub{
-		Method: "GET",
-		URL:    "/open-apis/wiki/v2/spaces/get_node",
-		Body: map[string]interface{}{
-			"code": 0, "msg": "ok",
-			"data": map[string]interface{}{
-				"node": map[string]interface{}{
-					"obj_type":  "docx",
-					"obj_token": "doxcnResolved123",
-				},
-			},
-		},
-	})
-
-	tmpDir := t.TempDir()
-	withDocsWorkingDir(t, tmpDir)
-
-	err := mountAndRunDocs(t, DocMediaInsert, []string{
-		"+media-insert",
-		"--doc", "https://example.larksuite.com/wiki/xxxxxx",
-		"--file", "missing.png",
-		"--as", "bot",
-	}, f, nil)
-	if err == nil {
-		t.Fatal("expected file-not-found error, got nil")
-	}
-	if !strings.Contains(err.Error(), "file not found") {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	if !strings.Contains(stderr.String(), "Resolved wiki to docx") {
-		t.Fatalf("stderr missing wiki resolution log: %s", stderr.String())
 	}
 }
